--- src/format/test/constructor-nested-function-spacing.test.ts
@@ -1,32 +1,0 @@
-import assert from "node:assert/strict";
-import { describe, it } from "node:test";
-
-import { Format } from "../src/index.js";
-
-void describe("constructor nested function spacing", () => {
-    void it("omits a blank line after nested functions before closing the constructor", async () => {
-        const source = [
-            "function Outer() constructor {",
-            "    function inner() {",
-            "        return 1;",
-            "    }",
-            "}",
-            ""
-        ].join("\n");
-
-        const formatted = await Format.format(source, {
-            parser: "gml-parse"
-        });
-
-        const lines = formatted.split("\n");
-        const closingBraceIndex = lines.lastIndexOf("}");
-
-        assert.ok(closingBraceIndex > 0, "Formatted constructor should include a closing brace.");
-
-        assert.notStrictEqual(
-            lines[closingBraceIndex - 1],
-            "",
-            "Formatter should not include a blank line between the nested function and the constructor closing brace."
-        );
-    });
-});
